requirements.txt
django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
bcrypt==4.2.1
psycopg2-binary==2.9.10